MIPS: build fix: sprinkle hardfloat into macros and embedded ASM
GCC before 4.9 haven't passed CFLAGS to the assembler properly,
in this case that flag is -msoft-float.
GCC 4.9 now passes all the CFLAGS to assembler as it should,
but that caused assembler errors in a lot of files:

"Error: opcode not supported on this processor"

This patch fixes it by explicitly setting the hardfloat
attribute in those files.

.set hardfloat

Squashed with bdf828255a2e [MIPS] Fix build issue with 4.9 GCC

Change-Id: I82bfcd4a0d677eebd0fd380e01f018d9bdbf3d34
Signed-off-by: Leonid Yegoshin <Leonid.Yegoshin@imgtec.com>
Reviewed-on: https://mipsia.review.mips.com/2610
Reviewed-by: Raghu Gandham <raghu.gandham@imgtec.com>
Tested-by: Raghu Gandham <raghu.gandham@imgtec.com>