[MIPS64] Handle additional O32 syscall args in 64bit kernel
The kernel should not access the user stack unless it has to.
This patch adds the number of parameters that each system call
takes (cf scall32-o32.S) and uses this information to control
how many access are made to the user stack.

This issue was exposed by the Android threads code which
unmaps the thread stack before calling exit(2).

This commit also includes a fix for build issue introduced by
the following commit:

    commit 2c6d7de579c2cfebbc9378e5209c641a93839f0a
    Author: Kees Cook <keescook@chromium.org>
    Date:   Tue Jun 10 15:45:09 2014 -0700

        MIPS: add seccomp syscall

        Wires up the new seccomp syscall.

Change-Id: Ie33adaf716a1855d1cb59c2f974cc44febbb41dc
Reviewed-on: https://mipsia.review.mips.com/2855
Reviewed-by: Raghu Gandham <raghu.gandham@imgtec.com>
Tested-by: Raghu Gandham <raghu.gandham@imgtec.com>