x86: Add ranchu defconfigs for the new Android emulator
Two new defconfigs, i386_ranchu and x86_64_ranchu, are created to build
kernels for use with the new Android emulator (qemu-android). They are
adapted from i386_emu and x86_64_emu, respectively. Changes include:

 * Enable booting from UEFI.
 * Enable virtio drivers.
 * Disable drivers for unused goldfish devices.
 * Disable MTD and MMC drivers.
 * Enable I/O APIC support. (i386 only)
 * Enable the serial port-based system console. (i386 only)

In addition, fix the Makefile for arch/x86/platform/goldfish/ so that
source files in this directory are excluded from the build if the
goldfish platform bus is disabled, which is the case with the ranchu
defconfigs.

Change-Id: I8a77ed8c37728e7b896959df7dfc7eca1342e128
Signed-off-by: Yu Ning <yu.ning@intel.com>